toiling on sha512

+++ constants/constants.go
@@ -23,8 +23,7 @@ var (
 	}
 )
 
-
-var Sha512K = [80]uint{
+var Sha512K = [80]uint64{
 	0x428a2f98d728ae22, 0x7137449123ef65cd,
 	0xb5c0fbcfec4d3b2f, 0xe9b5dba58189dbbc,
 	0x3956c25bf348b538, 0x59f111f1b605d019,
